+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Sin Publicar
 - Fix accessibility in carousel touchpoints
+- adding a full type in Touchpoints carousel
 
 # v1.34.1
 🚀 1.34.1 🚀